under_construction

+++ src/utils/utils.c
@@ -34,9 +34,9 @@ char *ft_strcat(char *dest, const char *src)
 	return (dest);
 }
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
